requirements.txt
# Data Governance Framework Dependencies

# Snowflake connector
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Data validation
great-expectations==0.18.8

# Visualization (for lineage graphs)
networkx==3.2.1
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
